solaris-userland: 16-test-vmlimit.patch
This patch for a Solaris-specific problem but is a good safety precaution,
so although it is not suitable for upstream as-is, we might offer it in a
slightly tweaked form at some point in the future.

Note that this can also have wrong inpact on other tests (e.g., imap which
tries to allocate huge chunks of virtual memory).