solaris-userland: 20-test_fixes.patch
This patch disables several tests that doesn't work correctly on Solaris.

test_gdb is not deterministic and sometimes fails for unimportant reasons (only
when ran with gmake test from component root). Same is true for on ofe the
test_normalization tests. [Not for upstream]

test_time changes used locale to region specific utf8 as python is not able to
work with non utf8 region specific encodings in Solaris. [Not for upstream].
Second change changes skip reason string to reflect the actual platform.
[Can be offered upstream]

test_socket fixes problem with sparc specific behavior of CMSG_SPACE(). Needed
values are not linear as the test expects and that might cause an exception for
last byte. Other problems in this test are related to known issues also
affecting other platforms. [Can be atleast reported upstream]

test_socket also disables testRecvmsgTrunc because MSG_TRUNC flag on Solaris is
currently broken (see 29193312) [Not for upstream].

test_structures is related to known bad support of reference passing on many
platforms. This just skips given test [Not for upstream].

test_re those two test are probably also locale related but I have no idea what
is the root cause of these... [Should be reported upstream]

test_float changes encodings for the same reason as in test_time.
[Not for upstream]

test_faulthandler is segfaulting on intel in this test (end of program with
registered faulthandler with chain=True). This bug was reported upstream.
[https://bugs.python.org/issue35484]

test_site is not deterministic and sometimes fails not sure why.
[Not for upstream]