solaris-userland: 18-obmalloc-adi.patch
This patch was developed in house for Bug 21658934. Python PyObject_Free()
implementation relies on being able to read memory that might not belong to
the current buffer. When ADIHEAP is enabled, this is detected as a violation.
Use an explicit nonfaulting load to ignore the ADI tag.

This patch should now work with both Studio and gcc compilers (and possibly
many others). It might be pushed upstream.