solaris-userland: 23-socket.patch
Sendfile on Solaris can raise EINVAL if offset is bigger than size of the file.
Because this would not send any data anyway, we can safely avoid calling this
function at all.

With this patch, one test must be skipped. It expects exception when remote
host disconnects during the transmision.  But as the test is right now it sends
all the data and than check for disconnect, which is never raised, because
connection is not checked when no data are there to send.

This should not be a problem and can be ignored.

This is not intended for upstream.