solaris-userland: 25-sparc-pgo.patch
At optimization levels above O0, gcc generates unusable code on SPARC when
profile guided optimization is enabled.  Disabling delay-slot instruction
scheduling avoids this problem.  Problem is currently being investigated.
(Bug. 29167790)

This is done in configure rather than in Makefile, because here we can
disable it in the profile generation phase and reenable in the final package.

This patch is not suitable for upstream.